refactor : plan 엔티티, 패키지 수정

+++ backend/src/main/java/com/ssafy/enjoytrip/core/plan/model/dto/PlanBoardDto.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 @Builder
 public class PlanBoardDto {
 
-    private int planBoardId;
+    private Long planBoardId;
     private String userId;
     private String title;
     private Date startDate;
